Insert is not on top of the original list
