+++ foods.html
@@ -15,6 +15,7 @@
 	    Calories:<br>
 	    <input id='calories-field' type="text" name="food-calories"><br><br>
 	    <input type="submit" id='food-submit' value="Submit Food">
+	    
 	  </fieldset>
 	</form>
 </div>
